+++ CHANGELOG.rst
@@ -5,9 +5,7 @@
   MemoFileNotFound. They all inherit from IOError, so old code will
   still work.
 
-* bugfix: numeric parser (N) failed when the data was a string. It now
-  returns an InvalidValue object, which is a byte string that contains
-  the raw data.
+* added InvalidValue.
 
 
 1.1.1 - 2014-08-03